apps/vision_tools/requirements.txt
fastapi>=0.125.0
uvicorn[standard]>=0.38.0
numpy>=2.0.0
pillow>=11.0.0
opencv-python-headless>=4.10.0

